Sync the garbage collection branch with 1.8.4p0

+++ src/sort.c
@@ -292,7 +292,7 @@ GENRECORD(gen_magic)
     case '8':
     case '9':
       intval = 0;
-      while (*s && isdigit(*s)) {
+      while (*s && isdigit((int) *s)) {
         intval *= 10;
         intval += *s - '0';
         s++;
@@ -300,10 +300,10 @@ GENRECORD(gen_magic)
       safe_format(buff, &bp, "%.20d", intval);
       if (*s == '.') {
         s++;
-        if (isdigit(*s)) {
+        if (isdigit((int) *s)) {
           intval = 0;
           numdigits = 0;
-          while (*s && isdigit(*s)) {
+          while (*s && isdigit((int) *s)) {
             intval *= 10;
             intval += *s - '0';
             numdigits++;
@@ -319,10 +319,10 @@ GENRECORD(gen_magic)
       }
       break;
     case NUMBER_TOKEN:
-      if (isdigit(*(s + 1))) {
+      if (isdigit((int) *(s + 1))) {
         s++;
         victim = 0;
-        while (*s && isdigit(*s)) {
+        while (*s && isdigit((int) *s)) {
           victim *= 10;
           victim += *s - '0';
           s++;
